+++ annotation-config.json
@@ -1,0 +1,10 @@
+{
+  "originMap": {
+    "http://cmdb.ft.com/systems/methode-web-pub": "annotations-v1",
+    "http://cmdb.ft.com/systems/next-video-editor": "annotations-next-video"
+  },
+  "lifecycleMap": {
+    "annotations-next-video": "next-video",
+    "annotations-v1": "v1"
+  }
+}
