+++ http_handler_test.go
@@ -31,6 +31,8 @@ type HttpHandlerTestSuite struct {
 	producer           *mockProducer
 	message            kafka.FTMessage
 	healthCheckHandler healthCheckHandler
+	originMap          map[string]string
+	lifecycleMap       map[string]string
 }
 
 func (suite *HttpHandlerTestSuite) SetupTest() {
@@ -49,6 +51,9 @@ func (suite *HttpHandlerTestSuite) SetupTest() {
 	assert.NoError(suite.T(), err, "Unexpected error")
 	suite.message = kafka.NewFTMessage(headers, string(msgBody))
 	suite.healthCheckHandler = healthCheckHandler{}
+	suite.originMap, suite.lifecycleMap = readConfigMap("annotation-config.json")
+
+	assert.NoError(suite.T(), err, "Unexpected error")
 }
 
 func TestHttpHandlerTestSuite(t *testing.T) {
@@ -58,66 +63,60 @@ func TestHttpHandlerTestSuite(t *testing.T) {
 func (suite *HttpHandlerTestSuite) TestPutHandler_Success() {
 	suite.annotationsService.On("Write", knownUUID, annotationLifecycle, platformVersion, suite.annotations).Return(nil)
 	suite.producer.On("SendMessage", mock.Anything).Return(nil).Once()
-	request := newRequest("PUT", fmt.Sprintf("/content/%s/annotations", knownUUID), "application/json", suite.body)
+	request := newRequest("PUT", fmt.Sprintf("/content/%s/annotations/%s", knownUUID, annotationLifecycle), "application/json", suite.body)
 	request.Header.Add("X-Request-Id", "tid_sample")
-	request.Header.Add("X-Origin-System-Id", "http://cmdb.ft.com/systems/methode-web-pub")
-	httpHandler := httpHandlers{AnnotationsService: suite.annotationsService, producer: suite.producer}
+	httpHandler := httpHandler{suite.annotationsService, suite.producer, suite.originMap, suite.lifecycleMap}
 	rec := httptest.NewRecorder()
-	router(httpHandler, suite.healthCheckHandler).ServeHTTP(rec, request)
+	router(&httpHandler, &suite.healthCheckHandler).ServeHTTP(rec, request)
 	assert.True(suite.T(), http.StatusCreated == rec.Code, fmt.Sprintf("Wrong response code, was %d, should be %d", rec.Code, http.StatusCreated))
 	assert.JSONEq(suite.T(), message("Annotations for content 12345 created"), rec.Body.String(), "Wrong body")
 	suite.producer.AssertExpectations(suite.T())
 }
 
 func (suite *HttpHandlerTestSuite) TestPutHandler_ParseError() {
-	request := newRequest("PUT", fmt.Sprintf("/content/%s/annotations", knownUUID), "application/json", []byte(`{"id": "1234"}`))
+	request := newRequest("PUT", fmt.Sprintf("/content/%s/annotations/%s", knownUUID, annotationLifecycle), "application/json", []byte(`{"id": "1234"}`))
 	request.Header.Add("X-Request-Id", "tid_sample")
-	request.Header.Add("X-Origin-System-Id", "http://cmdb.ft.com/systems/methode-web-pub")
-	httpHandler := httpHandlers{AnnotationsService: suite.annotationsService, producer: suite.producer}
+	httpHandler := httpHandler{suite.annotationsService, suite.producer, suite.originMap, suite.lifecycleMap}
 	rec := httptest.NewRecorder()
-	router(httpHandler, suite.healthCheckHandler).ServeHTTP(rec, request)
+	router(&httpHandler, &suite.healthCheckHandler).ServeHTTP(rec, request)
 	assert.True(suite.T(), http.StatusBadRequest == rec.Code, fmt.Sprintf("Wrong response code, was %d, should be %d", rec.Code, http.StatusBadRequest))
 }
 
 func (suite *HttpHandlerTestSuite) TestPutHandler_ValidationError() {
-	request := newRequest("PUT", fmt.Sprintf("/content/%s/annotations", knownUUID), "application/json", []byte(`"{"thing": {"prefLabel": "Apple"}`))
+	request := newRequest("PUT", fmt.Sprintf("/content/%s/annotations/%s", knownUUID, annotationLifecycle), "application/json", []byte(`"{"thing": {"prefLabel": "Apple"}`))
 	request.Header.Add("X-Request-Id", "tid_sample")
-	request.Header.Add("X-Origin-System-Id", "http://cmdb.ft.com/systems/methode-web-pub")
-	httpHandler := httpHandlers{AnnotationsService: suite.annotationsService, producer: suite.producer}
+	httpHandler := httpHandler{suite.annotationsService, suite.producer, suite.originMap, suite.lifecycleMap}
 	rec := httptest.NewRecorder()
-	router(httpHandler, suite.healthCheckHandler).ServeHTTP(rec, request)
+	router(&httpHandler, &suite.healthCheckHandler).ServeHTTP(rec, request)
 	assert.True(suite.T(), http.StatusBadRequest == rec.Code, fmt.Sprintf("Wrong response code, was %d, should be %d", rec.Code, http.StatusBadRequest))
 }
 
 func (suite *HttpHandlerTestSuite) TestPutHandler_NotJson() {
-	request := newRequest("PUT", fmt.Sprintf("/content/%s/annotations", knownUUID), "text/html", suite.body)
-	httpHandler := httpHandlers{AnnotationsService: suite.annotationsService, producer: suite.producer}
+	request := newRequest("PUT", fmt.Sprintf("/content/%s/annotations/%s", knownUUID, annotationLifecycle), "text/html", suite.body)
+	httpHandler := httpHandler{suite.annotationsService, suite.producer, suite.originMap, suite.lifecycleMap}
 	rec := httptest.NewRecorder()
-	router(httpHandler, suite.healthCheckHandler).ServeHTTP(rec, request)
+	router(&httpHandler, &suite.healthCheckHandler).ServeHTTP(rec, request)
 	assert.True(suite.T(), http.StatusBadRequest == rec.Code, fmt.Sprintf("Wrong response code, was %d, should be %d", rec.Code, http.StatusBadRequest))
-
 }
 
 func (suite *HttpHandlerTestSuite) TestPutHandler_WriteFailed() {
 	suite.annotationsService.On("Write", knownUUID, annotationLifecycle, platformVersion, suite.annotations).Return(errors.New("Write failed"))
-	request := newRequest("PUT", fmt.Sprintf("/content/%s/annotations", knownUUID), "application/json", suite.body)
+	request := newRequest("PUT", fmt.Sprintf("/content/%s/annotations/%s", knownUUID, annotationLifecycle), "application/json", suite.body)
 	request.Header.Add("X-Request-Id", "tid_sample")
-	request.Header.Add("X-Origin-System-Id", "http://cmdb.ft.com/systems/methode-web-pub")
-	httpHandler := httpHandlers{AnnotationsService: suite.annotationsService, producer: suite.producer}
+	httpHandler := httpHandler{suite.annotationsService, suite.producer, suite.originMap, suite.lifecycleMap}
 	rec := httptest.NewRecorder()
-	router(httpHandler, suite.healthCheckHandler).ServeHTTP(rec, request)
+	router(&httpHandler, &suite.healthCheckHandler).ServeHTTP(rec, request)
 	assert.True(suite.T(), http.StatusServiceUnavailable == rec.Code, fmt.Sprintf("Wrong response code, was %d, should be %d", rec.Code, http.StatusServiceUnavailable))
 }
 
 func (suite *HttpHandlerTestSuite) TestPutHandler_ForwardingFailed() {
 	suite.annotationsService.On("Write", knownUUID, annotationLifecycle, platformVersion, suite.annotations).Return(nil)
 	suite.producer.On("SendMessage", mock.Anything).Return(errors.New("Forwarding failed"))
-	request := newRequest("PUT", fmt.Sprintf("/content/%s/annotations", knownUUID), "application/json", suite.body)
+	request := newRequest("PUT", fmt.Sprintf("/content/%s/annotations/%s", knownUUID, annotationLifecycle), "application/json", suite.body)
 	request.Header.Add("X-Request-Id", "tid_sample")
-	request.Header.Add("X-Origin-System-Id", "http://cmdb.ft.com/systems/methode-web-pub")
-	httpHandler := httpHandlers{AnnotationsService: suite.annotationsService, producer: suite.producer}
+	httpHandler := httpHandler{suite.annotationsService, suite.producer, suite.originMap, suite.lifecycleMap}
 	rec := httptest.NewRecorder()
-	router(httpHandler, suite.healthCheckHandler).ServeHTTP(rec, request)
+	router(&httpHandler, &suite.healthCheckHandler).ServeHTTP(rec, request)
 	assert.True(suite.T(), http.StatusInternalServerError == rec.Code, fmt.Sprintf("Wrong response code, was %d, should be %d", rec.Code, http.StatusInternalServerError))
 }
 
@@ -125,7 +124,7 @@ func (suite *HttpHandlerTestSuite) TestGetHandler_Success() {
 	suite.annotationsService.On("Read", knownUUID, annotationLifecycle).Return(suite.annotations, true, nil)
 	request := newRequest("GET", fmt.Sprintf("/content/%s/annotations/%s", knownUUID, annotationLifecycle), "application/json", nil)
 	rec := httptest.NewRecorder()
-	router(httpHandlers{suite.annotationsService, suite.producer}, suite.healthCheckHandler).ServeHTTP(rec, request)
+	router(&httpHandler{suite.annotationsService, suite.producer, suite.originMap, suite.lifecycleMap}, &suite.healthCheckHandler).ServeHTTP(rec, request)
 	assert.True(suite.T(), http.StatusOK == rec.Code, fmt.Sprintf("Wrong response code, was %d, should be %d", rec.Code, http.StatusOK))
 	expectedResponse, err := json.Marshal(suite.annotations)
 	assert.NoError(suite.T(), err, "")
@@ -136,7 +135,7 @@ func (suite *HttpHandlerTestSuite) TestGetHandler_NotFound() {
 	suite.annotationsService.On("Read", knownUUID, annotationLifecycle).Return(nil, false, nil)
 	request := newRequest("GET", fmt.Sprintf("/content/%s/annotations/%s", knownUUID, annotationLifecycle), "application/json", nil)
 	rec := httptest.NewRecorder()
-	router(httpHandlers{suite.annotationsService, suite.producer}, suite.healthCheckHandler).ServeHTTP(rec, request)
+	router(&httpHandler{suite.annotationsService, suite.producer, suite.originMap, suite.lifecycleMap}, &suite.healthCheckHandler).ServeHTTP(rec, request)
 	assert.True(suite.T(), http.StatusNotFound == rec.Code, fmt.Sprintf("Wrong response code, was %d, should be %d", rec.Code, http.StatusNotFound))
 }
 
@@ -144,7 +143,7 @@ func (suite *HttpHandlerTestSuite) TestGetHandler_ReadError() {
 	suite.annotationsService.On("Read", knownUUID, annotationLifecycle).Return(nil, false, errors.New("Read error"))
 	request := newRequest("GET", fmt.Sprintf("/content/%s/annotations/%s", knownUUID, annotationLifecycle), "application/json", nil)
 	rec := httptest.NewRecorder()
-	router(httpHandlers{suite.annotationsService, suite.producer}, suite.healthCheckHandler).ServeHTTP(rec, request)
+	router(&httpHandler{suite.annotationsService, suite.producer, suite.originMap, suite.lifecycleMap}, &suite.healthCheckHandler).ServeHTTP(rec, request)
 	assert.True(suite.T(), http.StatusServiceUnavailable == rec.Code, fmt.Sprintf("Wrong response code, was %d, should be %d", rec.Code, http.StatusServiceUnavailable))
 }
 
@@ -152,7 +151,7 @@ func (suite *HttpHandlerTestSuite) TestDeleteHandler_Success() {
 	suite.annotationsService.On("Delete", knownUUID, annotationLifecycle).Return(true, nil)
 	request := newRequest("DELETE", fmt.Sprintf("/content/%s/annotations/%s", knownUUID, annotationLifecycle), "application/json", nil)
 	rec := httptest.NewRecorder()
-	router(httpHandlers{suite.annotationsService, suite.producer}, suite.healthCheckHandler).ServeHTTP(rec, request)
+	router(&httpHandler{suite.annotationsService, suite.producer, suite.originMap, suite.lifecycleMap}, &suite.healthCheckHandler).ServeHTTP(rec, request)
 	assert.True(suite.T(), http.StatusNoContent == rec.Code, fmt.Sprintf("Wrong response code, was %d, should be %d", rec.Code, http.StatusNoContent))
 }
 
@@ -160,7 +159,7 @@ func (suite *HttpHandlerTestSuite) TestDeleteHandler_NotFound() {
 	suite.annotationsService.On("Delete", knownUUID, annotationLifecycle).Return(false, nil)
 	request := newRequest("DELETE", fmt.Sprintf("/content/%s/annotations/%s", knownUUID, annotationLifecycle), "application/json", nil)
 	rec := httptest.NewRecorder()
-	router(httpHandlers{suite.annotationsService, suite.producer}, suite.healthCheckHandler).ServeHTTP(rec, request)
+	router(&httpHandler{suite.annotationsService, suite.producer, suite.originMap, suite.lifecycleMap}, &suite.healthCheckHandler).ServeHTTP(rec, request)
 	assert.True(suite.T(), http.StatusNotFound == rec.Code, fmt.Sprintf("Wrong response code, was %d, should be %d", rec.Code, http.StatusNotFound))
 }
 
@@ -168,7 +167,7 @@ func (suite *HttpHandlerTestSuite) TestDeleteHandler_DeleteError() {
 	suite.annotationsService.On("Delete", knownUUID, annotationLifecycle).Return(false, errors.New("Delete error"))
 	request := newRequest("DELETE", fmt.Sprintf("/content/%s/annotations/%s", knownUUID, annotationLifecycle), "application/json", nil)
 	rec := httptest.NewRecorder()
-	router(httpHandlers{suite.annotationsService, suite.producer}, suite.healthCheckHandler).ServeHTTP(rec, request)
+	router(&httpHandler{suite.annotationsService, suite.producer, suite.originMap, suite.lifecycleMap}, &suite.healthCheckHandler).ServeHTTP(rec, request)
 	assert.True(suite.T(), http.StatusServiceUnavailable == rec.Code, fmt.Sprintf("Wrong response code, was %d, should be %d", rec.Code, http.StatusServiceUnavailable))
 }
 
@@ -176,7 +175,7 @@ func (suite *HttpHandlerTestSuite) TestCount_Success() {
 	suite.annotationsService.On("Count", annotationLifecycle, platformVersion).Return(10, nil)
 	request := newRequest("GET", fmt.Sprintf("/content/annotations/%s/__count", annotationLifecycle), "application/json", nil)
 	rec := httptest.NewRecorder()
-	router(httpHandlers{suite.annotationsService, suite.producer}, suite.healthCheckHandler).ServeHTTP(rec, request)
+	router(&httpHandler{suite.annotationsService, suite.producer, suite.originMap, suite.lifecycleMap}, &suite.healthCheckHandler).ServeHTTP(rec, request)
 	assert.True(suite.T(), http.StatusOK == rec.Code, fmt.Sprintf("Wrong response code, was %d, should be %d", rec.Code, http.StatusOK))
 }
 
@@ -184,78 +183,10 @@ func (suite *HttpHandlerTestSuite) TestCount_CountError() {
 	suite.annotationsService.On("Count", annotationLifecycle, platformVersion).Return(0, errors.New("Count error"))
 	request := newRequest("GET", fmt.Sprintf("/content/annotations/%s/__count", annotationLifecycle), "application/json", nil)
 	rec := httptest.NewRecorder()
-	router(httpHandlers{suite.annotationsService, suite.producer}, suite.healthCheckHandler).ServeHTTP(rec, request)
+	handler := httpHandler{suite.annotationsService, suite.producer, suite.originMap, suite.lifecycleMap}
+	fmt.Printf("hey: %v", handler.lifecycleMap)
+	router(&handler, &suite.healthCheckHandler).ServeHTTP(rec, request)
 	assert.True(suite.T(), http.StatusServiceUnavailable == rec.Code, fmt.Sprintf("Wrong response code, was %d, should be %d", rec.Code, http.StatusServiceUnavailable))
-}
-
-type QueueHandlerTestSuite struct {
-	suite.Suite
-	headers            map[string]string
-	body               []byte
-	message            kafka.FTMessage
-	queueMessage       queueMessage
-	annotationsService *mockAnnotationsService
-	producer           *mockProducer
-}
-
-func (suite *QueueHandlerTestSuite) SetupTest() {
-	var err error
-	suite.headers = createHeader("tid_sample", "http://cmdb.ft.com/systems/methode-web-pub")
-	suite.body, err = ioutil.ReadFile("exampleAnnotationsMessage.json")
-	assert.NoError(suite.T(), err, "Unexpected error")
-	suite.message = kafka.NewFTMessage(suite.headers, string(suite.body))
-	err = json.Unmarshal(suite.body, &suite.queueMessage)
-	assert.NoError(suite.T(), err, "Unexpected error")
-	suite.annotationsService = new(mockAnnotationsService)
-	suite.producer = new(mockProducer)
-}
-
-func TestQueueHandlerTestSuite(t *testing.T) {
-	suite.Run(t, new(QueueHandlerTestSuite))
-}
-
-func (suite *QueueHandlerTestSuite) TestQueueHandler_Ingest() {
-	suite.annotationsService.On("Write", suite.queueMessage.UUID, annotationLifecycle, platformVersion, suite.queueMessage.Payload).Return(nil)
-	suite.producer.On("SendMessage", suite.message).Return(nil)
-
-	qh := &queueHandler{
-		AnnotationsService: suite.annotationsService,
-		consumer:           mockConsumer{message: suite.message},
-		producer:           suite.producer,
-	}
-	qh.Ingest()
-
-	suite.annotationsService.AssertCalled(suite.T(), "Write", suite.queueMessage.UUID, annotationLifecycle, platformVersion, suite.queueMessage.Payload)
-	suite.producer.AssertCalled(suite.T(), "SendMessage", suite.message)
-}
-
-func (suite *QueueHandlerTestSuite) TestQueueHandler_Ingest_ProducerNil() {
-	suite.annotationsService.On("Write", suite.queueMessage.UUID, annotationLifecycle, platformVersion, suite.queueMessage.Payload).Return(nil)
-
-	qh := queueHandler{
-		AnnotationsService: suite.annotationsService,
-		consumer:           mockConsumer{message: suite.message},
-		producer:           nil,
-	}
-	qh.Ingest()
-
-	suite.annotationsService.AssertCalled(suite.T(), "Write", suite.queueMessage.UUID, annotationLifecycle, platformVersion, suite.queueMessage.Payload)
-	suite.producer.AssertNumberOfCalls(suite.T(), "SendMessage", 0)
-}
-
-func (suite *QueueHandlerTestSuite) TestQueueHandler_Ingest_JsonError() {
-	body := "invalid json"
-	message := kafka.NewFTMessage(suite.headers, string(body))
-
-	qh := &queueHandler{
-		AnnotationsService: suite.annotationsService,
-		consumer:           mockConsumer{message: message},
-		producer:           suite.producer,
-	}
-	qh.Ingest()
-
-	suite.producer.AssertNumberOfCalls(suite.T(), "SendMessage", 0)
-	suite.annotationsService.AssertNumberOfCalls(suite.T(), "Write", 0)
 }
 
 func newRequest(method, url, contentType string, body []byte) *http.Request {
