+++ suggestion-config.json
@@ -1,0 +1,8 @@
+{
+  "originMap": {
+    "concept-suggestor": "annotations-v2",
+  },
+  "lifecycleMap": {
+    "annotations-v2": "v2"
+  }
+}
